GameDevelopment_7
Close to finishing the foliage, just need to make a rocky wall around the island. Also added mountain asset for the back drop for my level and fused multiple assets into one for a cave. Increased the height jump for the high jump item for the player to collect. And created an entrance for level 2 (currently leads into level 0 (debug mode) right now) when the player get close towards the end of the level and once they collected the item.