GameDevelopment_11
Finally figure out the inventory and gain the ability to jump when the item is collected. Still sorting out the save and load within the different levels, but so far my character's position, inventory and powerup is able to save and load on the same level. And I finally finished the scenery by combining rock meshes together and creating a wall. and added an invisible incase the player escapes and fall through the rock walls.