GameDevelopment_21
Probably one of the last updates for a while before it is due in on Friday. I have created a couple of widgets and a level that shows the main-menu, instructions, and exit menu. As well for baking the lighting in the cave level, creating an exit object to the exit menu and readjusting the dialogue which I ended up stuck on and adjusting the poolsize on the forest level. I will clean up any unnecessary data, assets, levels, etc, before handing it in on Friday.