GameDevelopment_23
Fixed up the cave level, by changing the lighting to be top-down and dim down the brightness. Also added a glow object within the begging of the cave level and the end. Combined the rock assets to create the cave wall. added rock meshes within the cave's ground to prevent the player falling through. Added invisible walls around the bridge area of the level so players won't fall off and get stuck underneath the bridge.

Also added new level, "tutorial level" using starting with a blank level and adding the lighting and atmosphere similar to the forest level. Added the ocean within this level and currently sculpting the landscape