GameDevelopment_27
Have finished? cinematic scene for the introduction for my game, which means placing the camera far off into the scene and placing a cube for it to focus on as it zooms on to the island, played around with the camera component which includes "Current Aperture" and "Current Focal Length". I have also edited within the curve editor so the animation slowly eases in towards the island. Have added a fade in and out from for the introduction scene and added a camera shake so it feels like the player is moving with the waves. Now I am sorting out the background of the main island giving it more dimension