GameDevelopment_32
Downloaded a new pack, as I needed a gate for the player to open. I will have to delete the other assets that will not be needed. Also bought a asset pack that contains keys which I only copy and pasted those in along with their materials and also added a glow effect to them. Got rid of the fog blueprints, no longer necessary. Right now I am trying to make a spline mesh for the fencing which will be around the gate, so I am pushing this now incase anything bad happens.