GameDevelopment_35
Not going to lie, the last couple of days were very rough for me regarding this project. However, I managed to fix the speed upgrade so it doesn't disappear when entering into a new level. And right now I am trying to figure out the dialogue, I am not sure if I really like this code that I used so I might try looking at a different tutorial and coming up with a different solution