GameDevelopment_39
Made a lot of progress, I am just hoping it isn't useless progress as I really wanted an ending level and cinematic for my game.
Right now I am trying to figure out why the cutscene is being skipped when I link the end-menu widget towards the end. It might have something to do with the data or I might have another blueprint linked up to it somewhere? Either way I might I have to quickly make a new and fresh one to see if that is the problem?