GameDevelopment_40
Fixed the ending menu, so it doesn't interrupt the sequence player at the ending level. Decided to delete the datafiles whenever the player quits out of the game, this is easier for the time being as I only have one week left. Changed some UI element, updated the keyboard image, fixed the text and some typos within the UI (and within game) and added a pause HUD menu as well. Now all I need to do is to export this game and I should be finished with the alpha.